Split user facing encode args from internals
